Remove popup reducers/actions. Add Popup component.
